add more docs, cleanup code.
Still kept copyright as // to be consistent and we can migrate them all at once later
