+++ fhirengine/src/main/java/com/google/fhirengine/db/impl/DbTypeConverters.kt
@@ -16,12 +16,23 @@ package com.google.fhirengine.db.impl
 import androidx.room.TypeConverter
 import org.hl7.fhir.r4.model.ResourceType
 
-object DbTypeConverters {
+/**
+ * Type converters for Room to persist ResourceType as a string.
+ * see: https://developer.android.com/training/data-storage/room/referencing-data
+ */
+internal object DbTypeConverters {
     private val resourceTypeLookup = ResourceType.values().associateBy { it.name }
+    /**
+     * Converts a [ResourceType] into a String to be persisted in the database. This allows us to
+     * save [ResourceType] into the database while keeping it as the real type in entities.
+     */
     @JvmStatic
     @TypeConverter
     fun typeToString(resourceType: ResourceType) = resourceType.name
 
+    /**
+     * Converts a String into a [ResourceType]. Called when a query returns a [ResourceType].
+     */
     @JvmStatic
     @TypeConverter
     fun stringToResourceType(data: String) = resourceTypeLookup[data]
