+++ fhirengine/src/main/java/com/google/fhirengine/db/impl/ResourceEntity.kt
@@ -28,9 +28,7 @@ import org.hl7.fhir.r4.model.ResourceType
         ]
 )
 internal data class ResourceEntity(
-        @PrimaryKey(
-                autoGenerate = true
-        )
+        @PrimaryKey(autoGenerate = true)
         val id: Long,
         val resourceType: ResourceType,
         val resourceId: String,
