+++ fhirengine/src/main/java/com/google/fhirengine/db/impl/RoomResourceDb.kt
@@ -14,18 +14,12 @@
 
 package com.google.fhirengine.db.impl
 
-import android.content.Context
-import android.database.sqlite.SQLiteConstraintException
 import androidx.room.*
 import androidx.room.Dao
 import ca.uhn.fhir.parser.IParser
-import com.google.fhirengine.db.ResourceAlreadyExistsInDbException
-import com.google.fhirengine.db.ResourceNotFoundInDbException
 import com.google.fhirengine.index.FhirIndexer
-import com.google.fhirengine.resource.ResourceUtils
 import org.hl7.fhir.r4.model.Resource
 import org.hl7.fhir.r4.model.ResourceType
-import javax.inject.Inject
 
 @Database(
         entities = [ResourceEntity::class, StringIndexEntity::class, ReferenceIndexEntity::class],
@@ -112,58 +106,4 @@ internal abstract class Dao {
             resourceId: String,
             resourceType: ResourceType
     ): String?
-}
-
-
-internal class DatabaseImpl @Inject constructor(
-        context: Context,
-        private val iParser: IParser,
-        fhirIndexer: FhirIndexer
-) : com.google.fhirengine.db.Database {
-    val db = Room
-            .inMemoryDatabaseBuilder(context, RoomResourceDb::class.java)
-            // TODO https://github.com/jingtang10/fhir-engine/issues/32
-            //  don't allow main thread queries
-            .allowMainThreadQueries()
-            .build()
-    val dao by lazy {
-        db.dao().also {
-            it.fhirIndexer = fhirIndexer
-            it.iParser = iParser
-        }
-    }
-
-    override fun <R : Resource> insert(resource: R) {
-        try {
-            dao.insert(resource)
-        } catch (constraintException: SQLiteConstraintException) {
-            throw ResourceAlreadyExistsInDbException(
-                    resource.resourceType.name,
-                    resource.id,
-                    constraintException
-            )
-        }
-    }
-
-    override fun <R : Resource> update(resource: R) {
-        dao.update(resource)
-    }
-
-    override fun <R : Resource> select(clazz: Class<R>, id: String): R {
-        val type = ResourceUtils.getResourceType(clazz)
-        return dao.getResource(
-                resourceId = id,
-                resourceType = type
-        )?.let {
-            iParser.parseResource(clazz, it)
-        } ?: throw ResourceNotFoundInDbException(type.name, id)
-    }
-
-    override fun <R : Resource> delete(clazz: Class<R>, id: String) {
-        val type = ResourceUtils.getResourceType(clazz)
-        dao.deleteResource(
-                resourceId = id,
-                resourceType = type
-        )
-    }
 }
